requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
firebase-admin==7.1.0
python-dotenv==1.1.1
pydantic==2.11.7
google-generativeai==0.3.2
